requirements.txt
flask
pyserial
